Separated the service into separate modules
