+++ Controllers/PatientController.js
@@ -1,4 +1,4 @@
-import Service from '../Service.js'
+import PatientService from '../Services/PatientService.js'
 import createPatientSchema from '../schema/createPatientSchema.js'
 import validate from '../utilities/validate.js'
 
@@ -7,7 +7,7 @@ class PatientController {
     async createPatient(req, res) {
         const valid = validate(req.body, createPatientSchema)
         if (valid) {
-            const newPatient = await Service.createPatient(req.body.name)
+            const newPatient = await PatientService.createPatient(req.body.name)
             res.json(newPatient);
         } else {
             res.status(400).send('Patient name must contain 2 or more characters')
@@ -15,7 +15,7 @@ class PatientController {
     }
 
     async getAllPatients(req, res) {
-        const allPatients = await Service.getAllPatients();
+        const allPatients = await PatientService.getAllPatients();
         res.json(allPatients);
     }
 }
