+++ Controllers/QueueController.js
@@ -1,13 +1,13 @@
-import Service from '../Service.js'
+import QueueService from '../Services/QueueService.js'
 class QueueController {
     async getCurrentInQueue(req, res, next) {
-        const currentInQueue = await Service.getCurrentInQueue();
+        const currentInQueue = await QueueService.getCurrentInQueue();
         res.json(currentInQueue);
     }
 
     async nextPatientInQueue(req, res, next) {
         await Service.deleteFirstFromQueue();
-        const currentInQueue = await Service.getCurrentInQueue();
+        const currentInQueue = await QueueService.getCurrentInQueue();
         res.json(currentInQueue);
     }
 }
