+++ Controllers/ResolutionController.js
@@ -1,4 +1,4 @@
-import Service from '../Service.js'
+import ResolutionService from '../Services/ResolutionService.js'
 import createResolutionSchema from '../schema/createResolutionSchema.js'
 import validate from '../utilities/validate.js'
 
@@ -6,18 +6,18 @@ class ResolutionController {
     async createResolution(req, res) {
         const valid = validate(req.body, createResolutionSchema)
         if (valid) {
-            const newResolution = await Service.createResolution(req.body.resolution);
+            const newResolution = await ResolutionService.createResolution(req.body.resolution);
             res.json(newResolution);
         } else {
             res.status(400).send('The resolution must be at least 10 characters and no more than 400 characters.')
         }
     }
     async getResolution(req, res) {
-        const resolution = await Service.getResolution(req.body.name);
+        const resolution = await ResolutionService.getResolution(req.body.name);
         res.json(resolution);
     }
     async deleteResolution(req, res) {
-        const deletedResolution = await Service.deleteResolution(req.params.key);
+        const deletedResolution = await ResolutionService.deleteResolution(req.params.key);
         res.json(deletedResolution);
 
     }
