+++ Services/QueueService.js
@@ -1,0 +1,16 @@
+import Database from '../database.js'
+class QueueService {
+
+    async deleteFirstFromQueue() {
+        const deletedFirst = await Database.deleteFirstFromQueue();
+        return deletedFirst;
+    }
+
+    async getCurrentInQueue() {
+        const currentInQueue = await Database.getCurrentInQueue();
+        return currentInQueue;
+    }
+}
+
+
+export default new QueueService();
